% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbbs.R
\name{bbbs}
\alias{bbbs}
\title{Package for analyzing benthic biomass and body size}
\description{
tbbs package contains data and code to analyze benthic biomass and body size off SW Taiwan.
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>, Jian-Xiang Liao <jianxiangliao@gmail.com >
}
