% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{env}
\alias{env}
\title{Environmental data}
\format{
A matrix includes coordinates, depth, bottom water hydrography, sediment characteristic, and internal tide model outputs of sampling sites
}
\description{
Environmental data from the Gaoping Submarine Canyon and the adjacent slope during OR1 1096, 1102, 1114, and 1126.
}
\details{
Temperature, salinity, dissolved oxygen concentration, fluorescence, and light transmission of the bottom water were measured by
a conductivity-temperature-depth (CTD) recorder (Sea-Bird SBE 911) and other attached sensors. Surface sediment grain sizes (non-carbonate fraction)
were measured by a laser diffraction particle size analyzer (Beckman Coulter LS13 320). Total organic carbon (TOC) and total nitrogen (TN) were
analyzed with a Flash EA 1112 elemental analyzer. Porosity was estimated from wet weight and dry weight of sediment and assuming the dry sediment
density of 2.65 g/cm3. Hourly bottom current velocity at each site was derived from a 3-D, hydrodynamic internal tide model from Chiou et al. (2011).
Based on the internal tide model, we calculated the hourly mean velocity of bottom currents and the duration for which bottom current velocity
exceeded 20 cm/s for 1 month preceding the sampling campaign to evaluate the possible disturbances of near-bottom currents on nematodes.
}
\references{
\itemize{Liao, J.-X., Wei, C.-L.*, Yasuhara, M. (2020) Species and functional diversity of deep-sea nematodes in a high energy submarine canyon. Frontiers in Marine Science. fmars.2020.00591}
\itemize{Liao, J.-X., Chen, G.-M., Chiou, M., Jan, S., Wei, C.-L.* (2017). Internal tides affect benthic community structure in an energetic submarine canyon off SW Taiwan. Deep Sea Research Part I: Oceanographic Research Papers. 125: 147-160}
}
\keyword{datasets}
