% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.method.R
\name{MLE.method}
\alias{MLE.method}
\title{Abundance Size Specturm with MLE estimate for b}
\usage{
MLE.method(x)
}
\arguments{
\item{x}{A vector of individual body masses}
}
\value{
NBSS slope, confidence interval of slope; fitted values and upper CI and lower CI; individual rank and upper CI and lower CI
}
\description{
This function is modified from \url{https://github.com/andrew-edwards/fitting-size-spectra/blob/master/code/recommend/recommend.r} for re-producing Fig 6 in Edwards et al. (2017) with ggplot2
}
\examples{
library(sizeSpectra)
MLE.method(mei$Size)
}
\references{
Edwards, A.M., Robinson, J.P.W., Plank, M.J., Baum, J.K., Blanchard, J.L., 2017. Testing and recommending methods for fitting size spectra to data. Methods Ecol Evol 8, 57–67. \url{https://doi.org/10.1111/2041-210X.12641}
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>, Jian-Xiang Liao <jianxiangliao@gmail.com >
}
