% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpr.R
\docType{data}
\name{mpr}
\alias{mpr}
\title{Taxon data for calculating benthic productivity}
\format{
A matrix includes conversion factors, lifestyle traits, and environmental information.
}
\description{
Taxon information contains conversion factors, lifestyle traits, and environmental information needed to calculate annual benthic productivity.
}
\details{
This matrix contains data require for calculating benthic productivity using \link[BenthicPro]{BenthicPB} and is modified from \link[BenthicPro]{BenthProdExampleData}.
More details see \link[BenthicPro]{BenthicPB} and Thomas Brey's "Virtual Handbook on Population Dynamics" (\url{http://www.thomas-brey.de/science/virtualhandbook/spreadsheets/index.html})
}
\references{
\itemize{Henrike Andresen and Thomas Brey (2018). BenthicPro: Benthic Energy Flow. R package version 1.0.}
\itemize{Brey, Thomas (2012) A multi-parameter artificial neural network model to estimate macrobenthic invertebrate productivity and production. Limnol. Oceanogr.: Methods 10, 2012, 581-589. DOI 10.4319/lom.2012.10.581}
}
\keyword{datasets}
